feat: refactored login/signup and updated route protection, 2.Updated JWT payload 3.I made a change on protect and authorizedRole in routes, 4. Improved access control in user controllers

+++ controller/userController.js
@@ -1,31 +1,6 @@
 import User from '../model/userModel.js';
   
 
-export const signUp= async (req, res) => {
-  try {
-    const { firstName, lastName, email, password} = req.body;
-    if (!firstName || !lastName || !email || !password) {
-      return res.status(400).json({ error: "All fields are mandatory" });
-    }
-    const existingUser = await User.findOne({ email });
-    if (existingUser) {
-      return res.status(409).json({ error: "Email already exists" });
-    }
-
-    const user = await User.create({ firstName, lastName, email, password: hashedPassword });
-
-    if(user){
-    res.status(201).json({ message: `User created successfully`, firstName: user.firstName, lastName: user.lastName, email: user.email}) ;
-    
-  } else{
-    res.status(404); throw new Error("User data is not valid")
- }
- } catch (error) {
-    res.status(500).json({ message: error["message"] });
-  }
-}
-
-
 export const getAllUsers = async (req, res) => {
   try {
     const users = await User.find();
@@ -45,6 +20,7 @@ export const getUserById = async (req, res) => {
   }
 };
 
+
 export const updateUserById = async (req, res) => {
   try {
     const user = await User.findByIdAndUpdate(
@@ -61,6 +37,9 @@ export const updateUserById = async (req, res) => {
 
 export const deleteUserById = async (req, res) => {
   try {
+    if(req.user.role === "admin"){
+      return res.status(403).json({message: "only admins can delete users"})
+    }
     const user = await User.findByIdAndDelete(req.params.id);
     if (!user) return res.status(404).json({ error: 'User not found' });
     res.status(200).json({ message: 'User deleted successfully' });
@@ -68,4 +47,3 @@ export const deleteUserById = async (req, res) => {
     res.status(500).json({message: error["message"]});
   }
 };
-
